requirements.txt
sounddevice
soundfile
numpy
requests
faster-whisper
python-dotenv